fix(view): fixed type in render_count_valid_subarray_sum_of_k

+++ src/views/view.rs
@@ -163,7 +163,7 @@ impl View {
     pub fn render_count_valid_subarray_sum_of_k(&self) {
         let nums: Vec<i32> = vec![1, -1, 5, -2, 3];
         let k: i32 = 3;
-        let min_len: i32 = 2;
+        let min_len: isize = 2;
         println!(
             "Count of valid subarrays with sum k: {}",
             count_valid_subarray_sum_of_k(&nums, k, min_len),
